+++ src/pylogctx/celery.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 import logging
 
 from celery import current_task
-from celery.app.task import Task
+from celery import Task
 
 from .core import context
 
@@ -12,7 +12,22 @@ logger = logging.getLogger(__name__)
 
 
 class LoggingTask(Task):
+
+    def before_call(self):
+        # Override if you need to add some vars or log something.
+        pass
+
+    def after_call(self):
+        # Override if you need to add some vars or log something.
+        pass
+
     def __call__(self, *args, **kwargs):
+        # Save context to put in back at the end. Useful when a task call
+        # another task to avoid context from another task.
+        self.save_context = context.as_dict()
+        context.clear()
+
+        self.before_call()
         task = current_task
 
         try:
@@ -23,4 +38,7 @@ class LoggingTask(Task):
         try:
             return super(LoggingTask, self).__call__(*args, **kwargs)
         finally:
+            self.after_call()
             context.clear()
+            if self.save_context:
+                context.push(**self.save_context)